encoder/requirements.txt
torch
pandas
numpy
psycopg2-binary
python-dotenv
